+++ lib/services/subscription_service.dart
@@ -1,19 +1,20 @@
 import 'package:dio/dio.dart';
 import 'dart:async';
-import'package:firebase_auth/firebase_auth.dart';
+import 'package:gfg_mobile/services/auth_service.dart';
 
 class SubscriptionService {
   SubscriptionService();
-  final FirebaseAuth auth = FirebaseAuth.instance;
   final dio = new Dio();
+  final auth = new AuthService();
 
+  // returns a future that resolves to a bool that indicates whether or not the user is subscribed to a specific player
   Future<bool> checkSubscription (String player) async {
     try {
-      final user = await auth.currentUser();
+      final uid = await auth.currentUser();
       Response response = await dio.post(
         "http://10.0.2.2:8080/amISubscribed",
         data: {
-          'uid': user.uid,
+          'uid': uid,
           'playerId': player
         },
       );
@@ -24,13 +25,14 @@ class SubscriptionService {
     }
   }
 
+  // returns a future that resolves to a bool that indicates whether or not a subscription was successfully added to Firebase Firestore
   Future<bool> addSubscription (String player, String playerName, String team, String teamName, String charityName, String charity) async {
     try {
-      final user = await auth.currentUser();
+      final uid = await auth.currentUser();
       Response response = await dio.post(
         "http://10.0.2.2:8080/subscribe",
         data: {
-          'uid': user.uid,
+          'uid': uid,
           'playerId': player,
           'name': playerName,
           'team': team,
@@ -47,13 +49,14 @@ class SubscriptionService {
     }
   }
 
+  // returns a promise that resolves to a bool that reports whether a subscription was successfully removed from Firebase Firestore
   Future<bool> removeSubscription (String player) async {
     try {
-      final user = await auth.currentUser();
+      final uid = await auth.currentUser();
       Response response = await dio.delete(
         'http://10.0.2.2:8080/unsubscribe',
         data: {
-          'uid': user.uid,
+          'uid': uid,
           'playerId': player
         }
       );
@@ -64,13 +67,14 @@ class SubscriptionService {
     }
   }
 
+  // returns a promise that resolves to a bool that reports whether a subscription was successfully removed from Firebase Firestore
   Future<bool> updateSubscription (String player, String charity, String charityId) async {
     try {
-      final user = await auth.currentUser();
+      final uid = await auth.currentUser();
       Response response = await dio.put(
         'http://10.0.2.2:8080/updateSubscription',
         data: {
-          'uid': user.uid,
+          'uid': uid,
           'playerId': player,
           'charityName': charity,
           'charityId': charityId,
@@ -83,10 +87,11 @@ class SubscriptionService {
     }
   }
 
+  // returns a promise that resovles to a List of subscriptions held by the currently logged in user
   Future<List> getSubscriptions () async {
     try {
-      final user = await auth.currentUser();
-      final url = 'http://10.0.2.2:8080/subscriptions' + user.uid;
+      final uid = await auth.currentUser();
+      final url = 'http://10.0.2.2:8080/subscriptions' + uid;
       Response response = await dio.get(url);
       return response.data;
     } catch (error) {