+++ lib/views/login_page.dart
@@ -19,15 +19,16 @@ enum FormType {
 
 class _LoginPageState extends State<LoginPage> {
 
-  final formKey = new GlobalKey<FormState>();
-  final validate = new UserValidator();
+  final formKey = new GlobalKey<FormState>(); // form key
+  final validate = new UserValidator(); // RegEx to make sure user email and password are valid
   final auth = new AuthService();
   final theme = new Themes();
 
   String _email;
   String _password;
-  FormType _form = FormType.login;
+  FormType _form = FormType.login; // default is login, only switch to register if the user taps that they don't have an account
 
+  // change to whichever form type is not currently being displayed
   void _formChange () async {
     setState(() {
       if (_form == FormType.register) {
@@ -38,65 +39,34 @@ class _LoginPageState extends State<LoginPage> {
     });
   }
 
-  void _createAccountPressed () async {
-    final form = formKey.currentState;
-    if (form.validate()) {
-      form.save();
-      auth.createAccount(_email, _password);
-      widget.onSignedIn();
-    } else {
-      print("the form was invalid somehow");
-      // TODO: add Widget for incorrect e-mail / pass?
-    }
-  }
-
-  void _loginPressed () async {
-    final form = formKey.currentState;
-    if (form.validate()) {
-      form.save();
-      auth.login(_email, _password);
-      widget.onSignedIn();
-    } else {
-      // either email or pass is invalid
-      // TODO: add Widget for incorrect e-mail / pass?
-    }
-  }
-
-  void _passwordReset () async {
-    final form = formKey.currentState;
-    form.save();
-    if (validate.isEmailAddressValid(_email)) {
-      auth.passwordReset(_email);
-    } else {
-      form.validate();
-    }
-  }
-
   @override
   Widget build(BuildContext context) {
     return new Scaffold(
       backgroundColor: theme.backgroundColor,
-      appBar: new AppBar(
-        title: new Text(
-          'Login',
-          style: theme.textStyle,
-        ),
-        centerTitle: true,
-      ),
+      appBar: _buildBar(context),
       body: new Container(
         padding: EdgeInsets.all(16.0),
-        child: new Form(
+        child: new Form( // handles the email and password input from users
           key: formKey,
           child: new Column(
             crossAxisAlignment: CrossAxisAlignment.stretch,
-            children: buildTextFields() + buildButtons(),
+            children: _buildTextFields() + _buildButtons(),
           ),
         )
       )
     );
   }
 
-  List<Widget> buildTextFields() {
+  Widget _buildBar(BuildContext context) {
+    return new AppBar(
+      title: theme.loginBarTitle,
+      centerTitle: true,
+    );
+
+  }
+
+  // returns the two text fields, one for email and one for password
+  List<Widget> _buildTextFields() {
     return [
       new TextFormField(
         style: theme.textStyle,
@@ -114,7 +84,8 @@ class _LoginPageState extends State<LoginPage> {
     ];
   }
 
-  List<Widget> buildButtons() {
+  // builds either the login or create account buttons as well as the transition button depending on which form is up
+  List<Widget> _buildButtons() {
     if (_form == FormType.login) {
       return [
         new RaisedButton(
@@ -144,6 +115,43 @@ class _LoginPageState extends State<LoginPage> {
           onPressed: _formChange,
         )
       ];
+    }
+  }
+
+  // handles call to backend to create account
+  void _createAccountPressed () async {
+    final form = formKey.currentState;
+    if (form.validate()) {
+      form.save();
+      auth.createAccount(_email, _password);
+      widget.onSignedIn();
+    } else {
+      print("the form was invalid somehow");
+      // TODO: add Widget for incorrect e-mail / pass?
+    }
+  }
+
+  // handles call to Firebase Auth to login
+  void _loginPressed () async {
+    final form = formKey.currentState;
+    if (form.validate()) {
+      form.save();
+      auth.login(_email, _password);
+      widget.onSignedIn();
+    } else {
+      // either email or pass is invalid
+      // TODO: add Widget for incorrect e-mail / pass?
+    }
+  }
+
+  // sends password reset to user
+  void _passwordReset () async {
+    final form = formKey.currentState;
+    form.save();
+    if (validate.isEmailAddressValid(_email)) {
+      auth.passwordReset(_email);
+    } else {
+      form.validate();
     }
   }
 }